expense_classifier_no_target/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
openpyxl==3.1.2
python-dotenv==1.0.0
openai==0.27.8
loguru==0.6.0
python-multipart==0.0.6
jinja2==3.1.2
tenacity==8.2.2




# fastapi==0.104.1
# uvicorn==0.24.0
# pandas==2.1.3
# openpyxl==3.1.2
# python-dotenv==1.0.0
# openai==0.27.8
# loguru==0.6.0
# python-multipart==0.0.6
# jinja2==3.1.2
